backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.0.1
beautifulsoup4==4.13.5
blinker==1.9.0
Brotli==1.1.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.2.18
EbookLib==0.19
einops==0.8.1
et_xmlfile==2.0.0
ezdxf==1.4.2
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.43.1
frozenlist==1.7.0
fsspec==2025.9.0
gdown==5.2.0
h11==0.16.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaldifst==1.7.17
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.2
llvmlite==0.45.0
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
networkx==3.5
ninja==1.13.0
numba==0.62.0
numpy==2.3.3
numpy-stl==3.2.0
omegaconf==2.3.0
onnxruntime==1.23.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
pdfkit==1.0.0
pikepdf==9.11.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
platformdirs==4.4.0
pooch==1.8.2
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydicom==3.0.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyMatting==1.1.14
pyparsing==3.2.5
PyPDF2==3.0.1
pypng==0.20220715.0
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
python-utils==3.9.1
pytube==15.0.0
pytz==2025.2
pyworld==0.3.5
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
regex==2025.9.18
rembg==2.0.67
reportlab==4.4.4
requests==2.32.5
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.47.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.9.20
tiktoken==0.11.0
tinycss2==1.4.0
torch==2.8.0
tqdm==4.67.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
webencodings==0.5.1
Werkzeug==3.1.3
wetext==0.1.0
wget==3.2
wrapt==1.17.3
xlsxwriter==3.2.9
yarl==1.20.1
yt-dlp==2025.8.22
